+++ CHANGES.rst
@@ -17,6 +17,8 @@
 - New features:
 
   - Add ``flycheck-tslint-args`` to pass additional arguments to tslint [GH-1186]
+  - Add an error explainer to the ``rpm-rpmlint`` checker using
+    ``rpmlint -I`` [GH-1235]
 
 - Improvements: